server/requirements.txt
Flask
pymongo
python-dotenv
flask-cors
gunicorn
Werkzeug
PyJWT
flask-cors